@@ -12,30 +12,24 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 class RegisterController extends ApiController
 {
-    /**
-     * @var SerializerInterface
-     */
-    private $serializer;
+    private SerializerInterface $serializer;
+    private ValidatorInterface $validator;
+    private CommandBus $bus;
 
-    /**
-     * @var ValidatorInterface
-     */
-    private $validator;
-
-    public function __construct(ValidatorInterface $validator, SerializerInterface $serializer)
+    public function __construct(CommandBus $bus, ValidatorInterface $validator, SerializerInterface $serializer)
     {
         $this->serializer = $serializer;
         $this->validator = $validator;
+        $this->bus = $bus;
     }
 
     /**
      * @Route("/register", name="register", methods={"POST"})
      *
      * @param Request $request
-     * @param CommandBus $bus
      * @return JsonResponse
      */
-    public function register(Request $request, CommandBus $bus): JsonResponse
+    public function register(Request $request): JsonResponse
     {
         $command = $this->serializer->deserialize($request->getContent(), UserCreateCommand::class, 'json');
         $violations = $this->validator->validate($command);
@@ -44,7 +38,7 @@ class RegisterController extends ApiController
             return $this->respondValidationErrors($violations);
         }
 
-        $bus->handle($command);
+        $this->bus->handle($command);
 
         return $this->respondWithSuccess('User successfully created');
     }
